requirements.txt
fastapi
uvicorn
qrcode
PyJWT
pandas
openpyxl
PyPDF2
Pillow